refactorization of a function
